// web/src/hooks/queries/gql-generated-uniswap.ts
export type Maybe<T> = T | null;
// ...
export type Bundle = {
  __typename?: "Bundle";
  ethPriceUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
};
// ...
export type Burn = {
  __typename?: "Burn";
  amount: Scalars["BigInt"]["output"];
  amount0: Scalars["BigDecimal"]["output"];
  amount1: Scalars["BigDecimal"]["output"];
  amountUSD?: Maybe<Scalars["BigDecimal"]["output"]>;
  id: Scalars["ID"]["output"];
  logIndex?: Maybe<Scalars["BigInt"]["output"]>;
  origin: Scalars["Bytes"]["output"];
  owner?: Maybe<Scalars["Bytes"]["output"]>;
  pool: Pool;
  tickLower: Scalars["BigInt"]["output"];
  tickUpper: Scalars["BigInt"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  token0: Token;
  token1: Token;
  transaction: Transaction;
};
// ...
export type Collect = {
  __typename?: "Collect";
  amount0: Scalars["BigDecimal"]["output"];
  amount1: Scalars["BigDecimal"]["output"];
  amountUSD?: Maybe<Scalars["BigDecimal"]["output"]>;
  id: Scalars["ID"]["output"];
  logIndex?: Maybe<Scalars["BigInt"]["output"]>;
  owner?: Maybe<Scalars["Bytes"]["output"]>;
  pool: Pool;
  tickLower: Scalars["BigInt"]["output"];
  tickUpper: Scalars["BigInt"]["output"];
  timestamp: Scalars["BigInt"]["output"];
  transaction: Transaction;
};
// ...
export type Factory = {
  __typename?: "Factory";
  id: Scalars["ID"]["output"];
  owner: Scalars["ID"]["output"];
  poolCount: Scalars["BigInt"]["output"];
  totalFeesETH: Scalars["BigDecimal"]["output"];
  totalFeesUSD: Scalars["BigDecimal"]["output"];
  totalValueLockedETH: Scalars["BigDecimal"]["output"];
  totalValueLockedETHUntracked: Scalars["BigDecimal"]["output"];
  totalValueLockedUSD: Scalars["BigDecimal"]["output"];
  totalValueLockedUSDUntracked: Scalars["BigDecimal"]["output"];
  totalVolumeETH: Scalars["BigDecimal"]["output"];
  totalVolumeUSD: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  untrackedVolumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export enum OrderDirection {
  Asc = "asc",
  Desc = "desc",
}

export type Pool = {
  __typename?: "Pool";
  burns: Array<Burn>;
  collectedFeesToken0: Scalars["BigDecimal"]["output"];
  collectedFeesToken1: Scalars["BigDecimal"]["output"];
  collectedFeesUSD: Scalars["BigDecimal"]["output"];
  collects: Array<Collect>;
  createdAtBlockNumber: Scalars["BigInt"]["output"];
  createdAtTimestamp: Scalars["BigInt"]["output"];
  feeGrowthGlobal0X128: Scalars["BigInt"]["output"];
  feeGrowthGlobal1X128: Scalars["BigInt"]["output"];
  feeTier: Scalars["BigInt"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidity: Scalars["BigInt"]["output"];
  liquidityProviderCount: Scalars["BigInt"]["output"];
  mints: Array<Mint>;
  observationIndex: Scalars["BigInt"]["output"];
  poolDayData: Array<PoolDayData>;
  poolHourData: Array<PoolHourData>;
  sqrtPrice: Scalars["BigInt"]["output"];
  swaps: Array<Swap>;
  tick?: Maybe<Scalars["BigInt"]["output"]>;
  ticks: Array<Tick>;
  token0: Token;
  token0Price: Scalars["BigDecimal"]["output"];
  token1: Token;
  token1Price: Scalars["BigDecimal"]["output"];
  totalValueLockedETH: Scalars["BigDecimal"]["output"];
  totalValueLockedToken0: Scalars["BigDecimal"]["output"];
  totalValueLockedToken1: Scalars["BigDecimal"]["output"];
  totalValueLockedUSD: Scalars["BigDecimal"]["output"];
  totalValueLockedUSDUntracked: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  untrackedVolumeUSD: Scalars["BigDecimal"]["output"];
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export type PoolHourData = {
  __typename?: "PoolHourData";
  close: Scalars["BigDecimal"]["output"];
  feeGrowthGlobal0X128: Scalars["BigInt"]["output"];
  feeGrowthGlobal1X128: Scalars["BigInt"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  high: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidity: Scalars["BigInt"]["output"];
  low: Scalars["BigDecimal"]["output"];
  open: Scalars["BigDecimal"]["output"];
  periodStartUnix: Scalars["Int"]["output"];
  pool: Pool;
  sqrtPrice: Scalars["BigInt"]["output"];
  tick?: Maybe<Scalars["BigInt"]["output"]>;
  token0Price: Scalars["BigDecimal"]["output"];
  token1Price: Scalars["BigDecimal"]["output"];
  tvlUSD: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export enum PoolHourData_OrderBy {
  Close = "close",
  FeeGrowthGlobal0X128 = "feeGrowthGlobal0X128",
  FeeGrowthGlobal1X128 = "feeGrowthGlobal1X128",
  FeesUsd = "feesUSD",
  High = "high",
  Id = "id",
  Liquidity = "liquidity",
  Low = "low",
  Open = "open",
  PeriodStartUnix = "periodStartUnix",
  Pool = "pool",
  PoolCollectedFeesToken0 = "pool__collectedFeesToken0",
  PoolCollectedFeesToken1 = "pool__collectedFeesToken1",
  PoolCollectedFeesUsd = "pool__collectedFeesUSD",
  PoolCreatedAtBlockNumber = "pool__createdAtBlockNumber",
  PoolCreatedAtTimestamp = "pool__createdAtTimestamp",
  PoolFeeGrowthGlobal0X128 = "pool__feeGrowthGlobal0X128",
  PoolFeeGrowthGlobal1X128 = "pool__feeGrowthGlobal1X128",
  PoolFeeTier = "pool__feeTier",
  PoolFeesUsd = "pool__feesUSD",
  PoolId = "pool__id",
  PoolLiquidity = "pool__liquidity",
  PoolLiquidityProviderCount = "pool__liquidityProviderCount",
  PoolObservationIndex = "pool__observationIndex",
  PoolSqrtPrice = "pool__sqrtPrice",
  PoolTick = "pool__tick",
  PoolToken0Price = "pool__token0Price",
  PoolToken1Price = "pool__token1Price",
  PoolTotalValueLockedEth = "pool__totalValueLockedETH",
  PoolTotalValueLockedToken0 = "pool__totalValueLockedToken0",
  PoolTotalValueLockedToken1 = "pool__totalValueLockedToken1",
  PoolTotalValueLockedUsd = "pool__totalValueLockedUSD",
  PoolTotalValueLockedUsdUntracked = "pool__totalValueLockedUSDUntracked",
  PoolTxCount = "pool__txCount",
  PoolUntrackedVolumeUsd = "pool__untrackedVolumeUSD",
  PoolVolumeToken0 = "pool__volumeToken0",
  PoolVolumeToken1 = "pool__volumeToken1",
  PoolVolumeUsd = "pool__volumeUSD",
  SqrtPrice = "sqrtPrice",
  Tick = "tick",
  Token0Price = "token0Price",
  Token1Price = "token1Price",
  TvlUsd = "tvlUSD",
  TxCount = "txCount",
  VolumeToken0 = "volumeToken0",
  VolumeToken1 = "volumeToken1",
  VolumeUsd = "volumeUSD",
}
// ...
export type Query = {
  __typename?: "Query";
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
  bundle?: Maybe<Bundle>;
  bundles: Array<Bundle>;
  burn?: Maybe<Burn>;
  burns: Array<Burn>;
  collect?: Maybe<Collect>;
  collects: Array<Collect>;
  factories: Array<Factory>;
  factory?: Maybe<Factory>;
  flash?: Maybe<Flash>;
  flashes: Array<Flash>;
  mint?: Maybe<Mint>;
  mints: Array<Mint>;
  pool?: Maybe<Pool>;
  poolDayData?: Maybe<PoolDayData>;
  poolDayDatas: Array<PoolDayData>;
  poolHourData?: Maybe<PoolHourData>;
  poolHourDatas: Array<PoolHourData>;
  pools: Array<Pool>;
  position?: Maybe<Position>;
  positionSnapshot?: Maybe<PositionSnapshot>;
  positionSnapshots: Array<PositionSnapshot>;
  positions: Array<Position>;
  swap?: Maybe<Swap>;
  swaps: Array<Swap>;
  tick?: Maybe<Tick>;
  tickDayData?: Maybe<TickDayData>;
  tickDayDatas: Array<TickDayData>;
  tickHourData?: Maybe<TickHourData>;
  tickHourDatas: Array<TickHourData>;
  ticks: Array<Tick>;
  token?: Maybe<Token>;
  tokenDayData?: Maybe<TokenDayData>;
  tokenDayDatas: Array<TokenDayData>;
  tokenHourData?: Maybe<TokenHourData>;
  tokenHourDatas: Array<TokenHourData>;
  tokens: Array<Token>;
  transaction?: Maybe<Transaction>;
  transactions: Array<Transaction>;
  uniswapDayData?: Maybe<UniswapDayData>;
  uniswapDayDatas: Array<UniswapDayData>;
};
// ...
export type Tick = {
  __typename?: "Tick";
  collectedFeesToken0: Scalars["BigDecimal"]["output"];
  collectedFeesToken1: Scalars["BigDecimal"]["output"];
  collectedFeesUSD: Scalars["BigDecimal"]["output"];
  createdAtBlockNumber: Scalars["BigInt"]["output"];
  createdAtTimestamp: Scalars["BigInt"]["output"];
  feeGrowthOutside0X128: Scalars["BigInt"]["output"];
  feeGrowthOutside1X128: Scalars["BigInt"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidityGross: Scalars["BigInt"]["output"];
  liquidityNet: Scalars["BigInt"]["output"];
  liquidityProviderCount: Scalars["BigInt"]["output"];
  pool: Pool;
  poolAddress?: Maybe<Scalars["String"]["output"]>;
  price0: Scalars["BigDecimal"]["output"];
  price1: Scalars["BigDecimal"]["output"];
  tickIdx: Scalars["BigInt"]["output"];
  untrackedVolumeUSD: Scalars["BigDecimal"]["output"];
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export type TickHourData = {
  __typename?: "TickHourData";
  feesUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  liquidityGross: Scalars["BigInt"]["output"];
  liquidityNet: Scalars["BigInt"]["output"];
  periodStartUnix: Scalars["Int"]["output"];
  pool: Pool;
  tick: Tick;
  volumeToken0: Scalars["BigDecimal"]["output"];
  volumeToken1: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
};
// ...
export type Token = {
  __typename?: "Token";
  decimals: Scalars["BigInt"]["output"];
  derivedETH: Scalars["BigDecimal"]["output"];
  feesUSD: Scalars["BigDecimal"]["output"];
  id: Scalars["ID"]["output"];
  name: Scalars["String"]["output"];
  poolCount: Scalars["BigInt"]["output"];
  symbol: Scalars["String"]["output"];
  tokenDayData: Array<TokenDayData>;
  totalSupply: Scalars["BigInt"]["output"];
  totalValueLocked: Scalars["BigDecimal"]["output"];
  totalValueLockedUSD: Scalars["BigDecimal"]["output"];
  totalValueLockedUSDUntracked: Scalars["BigDecimal"]["output"];
  txCount: Scalars["BigInt"]["output"];
  untrackedVolumeUSD: Scalars["BigDecimal"]["output"];
  volume: Scalars["BigDecimal"]["output"];
  volumeUSD: Scalars["BigDecimal"]["output"];
  whitelistPools: Array<Pool>;
};
// ...
export type _Meta_ = {
  __typename?: "_Meta_";
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   *
   */
  block: _Block_;
  /** The deployment ID */
  deployment: Scalars["String"]["output"];
  /** If `true`, the subgraph encountered indexing errors at some past block */
  hasIndexingErrors: Scalars["Boolean"]["output"];
};

export enum _SubgraphErrorPolicy_ {
  /** Data will be returned even if the subgraph has indexing errors */
  Allow = "allow",
  /** If the subgraph has indexing errors, data will be omitted. The default. */
  Deny = "deny",
}
// ...
export type GetPoolHourDatasQuery = {
  __typename?: "Query";
  poolHourDatas: Array<{
    __typename?: "PoolHourData";
    token0Price: string;
    token1Price: string;
    periodStartUnix: number;
    sqrtPrice: string;
    liquidity: string;
    pool: {
      __typename?: "Pool";
      id: string;
      liquidity: string;
      token0: { __typename?: "Token"; id: string; name: string };
      token1: { __typename?: "Token"; id: string; name: string };
    };
  }>;
};
// ...
export const GetPoolHourDatasDocument = gql`
    query GetPoolHourDatas($skip: Int = 0, $first: Int, $where: PoolHourData_filter, $orderBy: PoolHourData_orderBy, $orderDirection: OrderDirection, $block: Block_height, $subgraphError: _SubgraphErrorPolicy_! = deny) {
  poolHourDatas(
    skip: $skip
    first: $first
    orderBy: $orderBy
    orderDirection: $orderDirection
    where: $where
    block: $block
    subgraphError: $subgraphError
  ) {
    token0Price
    token1Price
    periodStartUnix
    sqrtPrice
    liquidity
    pool {
      id
      liquidity
      token0 {
        id
        name
      }
      token1 {
        id
        name
      }
    }
  }
}
    `;